Gemini-QA-ChatBot/requirements.txt
python-dotenv
streamlit
google-generativeai